local/st: apply multiple patches for st
The boxdraw patch renders customized lines/blocks/braille
characters for gapless alignment.

The anysize patch allows st to resize to any pixel size,
makes inner border dynamic and centers the content of the
terminal.

The hidecursor patch hides the X cursor whenever a key is
pressed and show it back when the mouse is moved in the
terminal.